+++ tests/testthat/test-run_beast2_from_options.R
@@ -241,7 +241,7 @@ test_that("Run with spaces in the input filename, for Windows", {
   )
 })
 
-test_that("Run with spaces in the ouput state filename, for Windows", {
+test_that("Run with spaces in the output state filename, for Windows", {
 
   if (!is_beast2_installed()) return()